Declare FindSpeed
Refactoring a bit more. Now FindSpeed is declared but not defined. This causes a linker error. Still red for testing.

The interface has been updated as we realize that a begin and end pointer are need. Passing an array is just a pointer to the beginning of the array, which is not enough information.