Implement TargetAreaToSpeed
This implementation is a bit more generalized and will be used by the more specific implementation at a later stage. 

This implementation copies much of the existing implementation, adds some white space and some parenthesis (!).

The failure tests are not failing. As is this test. Something is wrong with our algorithm.